fix: move UserActionStamps to its own file for Zeitwerk compatibility
Zeitwerk expects one top-level constant per file. Having UserActionStamps
defined in user_action_observer.rb caused flaky test failures when the
file wasn't autoloaded before the test ran.
