+++ app/models/user_action_stamps.rb
@@ -1,0 +1,19 @@
+# Concern to include in models that need created_by/updated_by tracking
+module UserActionStamps
+  extend ActiveSupport::Concern
+
+  included do
+    before_create :stamp_created_by
+    before_update :stamp_updated_by
+  end
+
+  private
+
+  def stamp_created_by
+    self.created_by_id = UserActionObserver.current_user&.id if respond_to?(:created_by_id=)
+  end
+
+  def stamp_updated_by
+    self.updated_by_id = UserActionObserver.current_user&.id if respond_to?(:updated_by_id=)
+  end
+end
